+++ src/components/App/App.tsx
@@ -1,17 +1,7 @@
 import React from 'react';
 import { Switch, Route, Redirect } from 'react-router-dom';
-import {
-  AuthProvider,
-  PrivateRoute,
-  Signup,
-  Signin,
-  Game,
-  Header,
-  Landing,
-  Settings,
-  NotFound,
-  Leaderboard,
-} from '../';
+import { PrivateRoute, Signup, Signin, Game, Header, Landing, Settings, NotFound, Leaderboard } from '../';
+import { AuthProvider } from '../../services/auth';
 import './App.scss';
 
 const App = () => {
