+++ src/services/Auth/helpers/withAuth.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
 import { Consumer } from '../AuthProvider';
-import { authProps } from '../types';
+import { AuthProps } from '../types';
 
-function withAuth<P extends authProps>(WrappedComponent: React.ComponentType<P>) {
-  return class extends React.Component<Omit<P, keyof authProps>> {
+function withAuth<P extends AuthProps>(WrappedComponent: React.ComponentType<P>) {
+  return class extends React.Component<Omit<P, keyof AuthProps>> {
     static displayName = `withAuth(${WrappedComponent.displayName || WrappedComponent.name})`;
     render() {
       return <Consumer>{(value) => <WrappedComponent {...value} {...(this.props as P)} />}</Consumer>;
