+++ src/services/Auth/index.ts
@@ -1,3 +1,3 @@
 export { AuthProvider } from './AuthProvider';
 export { withAuth } from './helpers/withAuth';
-export { authProps } from './types';
+export { AuthProps } from './types';
