+++ src/services/Auth/types.ts
@@ -1,8 +1,8 @@
-type authProps = {
+type AuthProps = {
   isAuthorized: boolean;
   authorize(): void;
   logout(): void;
   isLoad: boolean;
 };
 
-export { authProps };
+export { AuthProps };
